🤖 add new project structure with some cleaning and refactoring. still a lot to do.
