+++ src/app.js
@@ -4,28 +4,17 @@ const token = process.env.token;
 const Discord = require("discord.js");
 const client = new Discord.Client();
 const Ytdl = require("ytdl-core");
-const mongoose = require("mongoose");
 const connectDB = require("./data/db");
 const commandsController = require("./controllers/commands");
 const index = require("./routes/index");
 const cors = require("cors");
 app.use(cors());
 const logger = require("morgan");
-const bodyParser = require("body-parser");
-const axios = require("axios");
-const fs = require("fs");
-const path = require("path");
-const { collection } = require("./models/command");
-let commands = {};
-let ready = false;
 app.use(logger("dev"));
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
+app.use(express.urlencoded({ extended: true }));
+app.use(express.json());
 
-async function getCommands() {
-  commands = await commandsController.getCommands();
-  return commands;
-}
+let ready = false;
 const servers = {
   server: {
     connection: null,
@@ -33,206 +22,16 @@ const servers = {
   },
 };
 require("discord-buttons")(client);
-// client.on("guildMemberAdd", async (member) => {
-//   const channel = member.guild.channels.cache.find(
-//     (ch) => ch.id === "402972272870162435"
-//   );
-//   if (!channel) return;
 
-//   await channel.send(`Bem-vindo(a), ${member}! Isso aqui tá uma loucura, né?`, {
-//     files: [
-//       "https://media.tenor.com/images/799ed0a2c8870ebcceee8f61742a129d/tenor.gif?itemid=7911935!!!",
-//     ],
-//   });
-// });
-client.on("voiceStateUpdate", async (oldMember, newMember) => {
-  // just a temporary joke; still a lot to refactor
-  if (newMember.channelID === process.env.JOKE_CHANNEL_ID) {
-    
-    if (newMember.member.voice.channel) {
-      newMember.member.voice.channel.join();
-      servers.server.connection = await newMember.member.voice.channel.join()
+const messagesHandler = require("./events/message/message");
+const guildMemberAdd = require("./events/guildMemberAdd/guildMemberAdd");
+const voiceStateUpdate = require("./events/voiceStateUpdate/voiceStateUpdate")
+client.on("message", messagesHandler);
+client.on("guildMemberAdd", guildMemberAdd);
+client.on("voiceStateUpdate", voiceStateUpdate);
 
-      const dispatcher = servers.server.connection.play(
-        Ytdl(process.env.JOKE, {
-          filter: "audioonly",
-          quality: "highestaudio",
-          requestOptions: {
-            headers: {
-              cookie: process.env.COOKIE,
-            },
-          },
-        })
-      );
-
-      dispatcher.on("start", () => {
-        console.log("audio's now playing");
-      });
-      dispatcher.on("finish", () => {
-        newMember.member.voice.channel.leave()
-        ready = false;
-  
-      });
-      dispatcher.on("error", console.error);
-    }
-  }
-});
 client.on("message", async (msg) => {
-  await getCommands();
-
-  if (msg.content === "!comandos") {
-    const commandsList = await commands.map(
-      (ObjCommand) => ` !${ObjCommand.command}`
-    );
-    await msg.channel.send(
-      `Total: ${commandsList.length} comandos. ${commandsList}.`
-    );
-  }
-  if (msg.content === "!gato") {
-    axios
-      .get("https://api.thecatapi.com/v1/images/search")
-      .then((res) => {
-        msg.channel.send(res.data[0].url);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }
-  if (msg.content === "!fala") {
-    msg.channel.send("Meow! I'm a baby loli kitten!", {
-      tts: true,
-    });
-  }
-  if (msg.content === "user") {
-    console.log(msg.guild.systemChannel.id);
-  }
-
-  if (msg.content.startsWith("!kof")) {
-    let embed = null;
-    if (!msg.mentions.users.size) {
-      return msg.channel.send(
-        "Por favor, use o comando !kof marcando alguém. Ex.: !kof @leticia"
-      );
-    }
-    const player1 = msg.author.username.toUpperCase();
-    const player2 = msg.mentions.users.first().username.toUpperCase();
-    let countP1 = 0;
-    let countP2 = 0;
-    embed = new Discord.MessageEmbed()
-      .setTitle("Hora do show!")
-      .setDescription(`FT: ${player1} vs ${player2}`)
-      .setColor("BLUE")
-      .setFooter("Adicione ou remova 1 ponto clicando nos botões.")
-      .setImage(
-        "https://cdn.discordapp.com/attachments/402972272870162435/860613813170077736/kofcollection.gif"
-      );
-    const { MessageButton, MessageActionRow } = require("discord-buttons");
-    let button1Plus = new MessageButton()
-      .setStyle("green")
-      .setLabel("+1")
-      .setID("p1_plus_1");
-
-    let button1Minus = new MessageButton()
-      .setStyle("red")
-      .setLabel("-1")
-      .setID("p1_minus_1");
-
-    let button2Plus = new MessageButton()
-      .setStyle("green")
-      .setLabel("+1")
-      .setID("p2_plus_1");
-
-    let button2Minus = new MessageButton()
-      .setStyle("red")
-      .setLabel("-1")
-      .setID("p2_minus_1");
-
-    let buttonEnds = new MessageButton()
-      .setStyle("blurple")
-      .setLabel("Parar")
-      .setID("stop");
-
-    const buttons1 = new MessageActionRow()
-      .addComponent(button1Plus)
-      .addComponent(button2Plus);
-
-    const buttons2 = new MessageActionRow()
-      .addComponent(button1Minus)
-      .addComponent(button2Minus);
-
-    const button3 = new MessageActionRow().addComponent(buttonEnds);
-
-    let m = await msg.channel.send("Placar:", {
-      embed,
-      components: [buttons1, buttons2, button3],
-    });
-
-    const filter = (button) => button;
-    const collector = m.createButtonCollector(filter, { time: 60000 * 120 });
-    collector.on("collect", async (button) => {
-      button.defer();
-      if (button.id === "p1_plus_1") {
-        countP1 = countP1 + 1;
-        let embed1 = new Discord.MessageEmbed()
-          .setTitle(`${player1}: ${countP1} vs ${player2}: ${countP2}`)
-          .setFooter("Adicione ou remova 1 ponto clicando nos botões.");
-        await button.message.edit({
-          embed: embed1,
-          components: [buttons1, buttons2, button3],
-        });
-      } else if (button.id === "p1_minus_1") {
-        if (countP1 > 0) countP1 = countP1 - 1;
-
-        let embed2 = new Discord.MessageEmbed()
-          .setTitle(`${player1}: ${countP1} vs ${player2}: ${countP2}`)
-          .setFooter("Adicione ou remova 1 ponto clicando nos botões.");
-
-        await button.message.edit({
-          embed: embed2,
-          components: [buttons1, buttons2, button3],
-        });
-      } else if (button.id === "p2_plus_1") {
-        countP2 = countP2 + 1;
-
-        let embed3 = new Discord.MessageEmbed()
-          .setTitle(`${player1}: ${countP1} vs ${player2}: ${countP2}`)
-          .setFooter("Adicione ou remova 1 ponto clicando nos botões.");
-
-        await button.message.edit({
-          embed: embed3,
-          components: [buttons1, buttons2, button3],
-        });
-      } else if (button.id === "p2_minus_1") {
-        if (countP2 > 0) countP2 = countP2 - 1;
-
-        let embed4 = new Discord.MessageEmbed()
-          .setTitle(`${player1}: ${countP1} vs ${player2}: ${countP2}`)
-          .setFooter("Adicione ou remova 1 ponto clicando nos botões.");
-
-        await button.message.edit({
-          embed: embed4,
-          components: [buttons1, buttons2, button3],
-        });
-      } else if (button.id === "stop") {
-        let embed5 = new Discord.MessageEmbed()
-          .setTitle(`${player1}: ${countP1} vs ${player2}: ${countP2}`)
-          .setFooter(
-            `Resultado final: ${
-              countP1 === countP2
-                ? "ninguém venceu 🤭"
-                : countP1 > countP2
-                ? player1 + "🏆"
-                : player2 + "🏆"
-            }`
-          );
-        collector.stop();
-        await button.message.edit({
-          embed: embed5,
-        });
-      }
-    });
-  }
-  // this is temporary
+  // this is temporary ONLY FOR DATABASE COMMANDS!
   let splitCmd = "";
   if (msg.content.match(/^!/)) {
     splitCmd = msg.content.slice(1, msg.content.length).split(" ");
