+++ src/events/message/commands/fala.js
@@ -1,0 +1,5 @@
+module.exports = async (msg) => {
+  return await msg.channel.send("Meow! I'm a baby loli kitten! Just kiddin' I'm a bot!", {
+    tts: true,
+  });
+};
