+++ src/events/message/commands/tests.js
@@ -1,0 +1,4 @@
+module.exports = async (msg) => {
+  console.log(msg.guild.systemChannel.id);
+  return await msg.channel.send("tests!!!!");
+};
